+++ app/src/main/java/com/shashank/applaunch/remote/model/Minutely.kt
@@ -1,6 +1,6 @@
 package com.shashank.applaunch.remote.model
 
 data class Minutely(
-    val dt: Int,
-    val precipitation: Int
+    val dt: Float,
+    val precipitation: Float
 )
