+++ app/src/main/java/com/shashank/applaunch/remote/model/Weather.kt
@@ -8,5 +8,5 @@ data class Weather(
     val lon: Double,
     val minutely: List<Minutely>,
     val timezone: String,
-    val timezone_offset: Int
+    val timezone_offset: Float
 )
